pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydromace"
version = "0.2"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    {name = "Rokas Elijošius", email = "re344@cam.ac.uk"}
]
dependencies = [
    "mace-torch>=0.3.3",
    "typer>=0.12"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["hydromace"]
